pyproject.toml
[project]
name = "chorus"
version = "0.1.0"
description = "Production-grade chorus LLM inference framework"
requires-python = ">=3.9"

dependencies = [
    "vllm>=0.6.0",
    "torch>=2.0.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",
    "psutil>=5.9.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "plotly>=5.18.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
chorus-server = "chorus.server.grpc_server:main"
chorus-bench = "chorus.benchmarks.runner:main"

[build-system]
requires = ["setuptools>=61.0", "wheel", "grpcio-tools>=1.60.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["chorus", "chorus.core", "chorus.server", "chorus.server.protos", "chorus.client", "chorus.aggregation", "chorus.benchmarks", "chorus.dashboard"]

